Update MMM to use sql server db not h2
