+++ data-query/src/test/java/gov/va/api/health/dataquery/tools/DatamartExporter.java
@@ -1,9 +1,5 @@
 package gov.va.api.health.dataquery.tools;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.google.common.collect.ImmutableMap;
-import com.microsoft.sqlserver.jdbc.SQLServerDataSource;
 import gov.va.api.health.dataquery.service.controller.allergyintolerance.AllergyIntoleranceEntity;
 import gov.va.api.health.dataquery.service.controller.condition.ConditionEntity;
 import gov.va.api.health.dataquery.service.controller.diagnosticreport.DiagnosticReportCrossEntity;
@@ -16,18 +12,10 @@ import gov.va.api.health.dataquery.service.controller.observation.ObservationEnt
 import gov.va.api.health.dataquery.service.controller.patient.PatientEntity;
 import gov.va.api.health.dataquery.service.controller.patient.PatientSearchEntity;
 import gov.va.api.health.dataquery.service.controller.procedure.ProcedureEntity;
-import java.io.FileInputStream;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Properties;
-import java.util.function.Supplier;
 import javax.persistence.EntityManager;
-import javax.persistence.spi.PersistenceUnitInfo;
-import javax.sql.DataSource;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.cfg.AvailableSettings;
-import org.hibernate.jpa.HibernatePersistenceProvider;
 
 /**
  * This application will copy data out of the Mitre database into a local H2 database. It expects
@@ -62,7 +50,7 @@ public class DatamartExporter {
   EntityManager mitre;
 
   public DatamartExporter(String configFile, String outputFile) {
-    mitre = new Mitre(configFile).get();
+    mitre = new SqlServerDb(configFile, MANAGED_CLASSES).get();
     h2 = new LocalH2(outputFile, MANAGED_CLASSES).get();
   }
 
@@ -97,61 +85,5 @@ public class DatamartExporter {
               log.info("{}", e);
               h2.persist(e);
             });
-  }
-
-  private static class Mitre implements Supplier<EntityManager> {
-
-    private final Properties config;
-
-    @SneakyThrows
-    Mitre(String configFile) {
-      log.info("Loading Mitre connection configuration from {}", configFile);
-      config = new Properties(System.getProperties());
-      try (FileInputStream inputStream = new FileInputStream(configFile)) {
-        config.load(inputStream);
-      }
-    }
-
-    @Override
-    public EntityManager get() {
-
-      PersistenceUnitInfo info =
-          PersistenceUnit.builder()
-              .persistenceUnitName("mitre")
-              .jtaDataSource(mitreDataSource())
-              .managedClasses(MANAGED_CLASSES)
-              .properties(mitreProperties())
-              .build();
-      return new HibernatePersistenceProvider()
-          .createContainerEntityManagerFactory(
-              info,
-              ImmutableMap.of(
-                  AvailableSettings.JPA_JDBC_DRIVER,
-                  "com.microsoft.sqlserver.jdbc.SQLServerDriver"))
-          .createEntityManager();
-    }
-
-    DataSource mitreDataSource() {
-      SQLServerDataSource ds = new SQLServerDataSource();
-      ds.setUser(valueOf("spring.datasource.username"));
-      ds.setPassword(valueOf("spring.datasource.password"));
-      ds.setURL(valueOf("spring.datasource.url"));
-      return ds;
-    }
-
-    Properties mitreProperties() {
-      Properties properties = new Properties();
-      properties.put("hibernate.hbm2ddl.auto", "none");
-      properties.put("hibernate.show_sql", "false"); // <---- CHANGE TO TRUE TO DEBUG
-      properties.put("hibernate.format_sql", "true");
-      properties.put("hibernate.globally_quoted_identifiers", "true");
-      return properties;
-    }
-
-    private String valueOf(String name) {
-      String value = config.getProperty(name, "");
-      assertThat(value).withFailMessage("System property %s must be specified.", name).isNotBlank();
-      return value;
-    }
   }
 }
